Add workflow to hide tests/samples
This workflow will rename 'Tests' to 'Tests~', 'Samples' to 'Samples~',
and remove the associated .meta files. It will then commit and
force-push the result to a branch named the same as the original plus a
`-release` prefix.

This will allow Tests and Samples to remain available in a dev-oriented
project, while having a branch excluding these which can be pulled into
the package manager.

This workflow will apply to any branch which contains the workflow file.